+++ WDL/mergesort.h
@@ -23,41 +23,41 @@ static void WDL_mergesort(void *base, size_t nmemb, size_t size, int (*compar)(c
 
   do
   {
-	  if (compar(b1, b2) > 0) // out of order, go to full merge
-	  {
-      int sofar = b1-(char*)base;
+    if (compar(b1, b2) > 0) // out of order, go to full merge
+    {
+      size_t sofar = b1-(char*)base;
       memcpy(tmpspace,base,sofar);
-	    memcpy(tmpspace+sofar, b2, size);
+      memcpy(tmpspace+sofar, b2, size);
       b2 += size;
-	    n2--;
+      n2--;
 
-	    char *writeptr=tmpspace+sofar+size;
+      char *writeptr=tmpspace+sofar+size;
       while (n1 > 0 && n2 > 0)
       {
-	      if (compar(b1, b2) > 0)
-	      {
-	        memcpy(writeptr, b2, size);
-	        b2 += size;
-	        n2--;
-	      }
-	      else
-	      {
+        if (compar(b1, b2) > 0)
+        {
+          memcpy(writeptr, b2, size);
+          b2 += size;
+          n2--;
+        }
+        else
+        {
           memcpy(writeptr, b1, size);
           b1 += size;
-	        n1--;
-	      }
-  	    writeptr += size;
+          n1--;
+        }
+        writeptr += size;
       }
 
       if (n1 > 0) memcpy(writeptr, b1, n1 * size);
       memcpy(base, tmpspace, (nmemb - n2) * size);
 
       break;
-	  }
+    }
 
     // in order, just advance
     b1 += size;
-	  n1--;
+    n1--;
   }
   while (n1 > 0 && n2 > 0);
 }
